started simplification of creating instruments
